change confirm-password structure
