+++ src/App.tsx
@@ -3,17 +3,17 @@ import './App.scss';
 import { Route, Switch, Redirect } from 'react-router-dom';
 
 import * as datasets from "./Config/datasets"
-
 import LoadingView from './Components/Loading/LoadingView';
 import useLogin from './hook/custom/useLogin';
 import { useAppUserState } from './hook/useAppUser';
 import useInitializeApp from './hook/custom/useInitializeApp';
-const MapLayout = lazy(() => import('./Components/Map/MapLayout'));
 const LoginRoute = lazy(() => import('./routes/login'));
 const SignUpRoute = lazy(() => import('./routes/sign-up'));
 const ResetPasswordRoute = lazy(() => import('./routes/reset-password'));
+const ConfirmPasswordRoute = lazy(() => import('./routes/confirm-password'));
+const MapLayout = lazy(() => import('./Components/Map/MapLayout'));
+
 const Unauthorized = lazy(() => import('./Components/Unauthorized/Unauthorized'));
-const ConfirmPasswordView = lazy(() => import('./Components/ConfirmPassword/ConfirmPasswordView'));
 const DetailedPageView = lazy(() => import('./Components/DetailedPage/DetailedPageView'));
 const ProfileView = lazy(() => import('./Components/Profile/ProfileView'));
 const WorkPlan = lazy(() => import('./Components/Work/Plan/WorkPlan'));
@@ -27,14 +27,14 @@ const App = () => {
 
   useInitializeApp();
 
-  return <>{
+  return (
     <Switch>
       <Suspense fallback={<div>...</div>}>
         <Route path={`/login`} component={LoginRoute} />
         <Route path={`/sign-up`} component={SignUpRoute} />
-        <Route path={'/404'} component={Unauthorized} />
         <Route path={`/reset-password`} component={ResetPasswordRoute} />
-        <Route path={`/confirm-password`} component={ConfirmPasswordView} />
+        <Route path={`/confirm-password`} component={ConfirmPasswordRoute} />
+        <Route path={'/404'} component={Unauthorized} />
         <Route path={`/detailed-page`} component={DetailedPageView} />
         <Route exact path="/" render={() => (
           <Redirect to="/login" />
@@ -51,8 +51,8 @@ const App = () => {
           appUser.designation === 'staff') && (appUser.status === 'approved') && <Route path={`/upload-attachment`} component={UploadAttachmentView} />}
         {(loading && <Route path={`/`} component={LoadingView} />)}
       </Suspense>
-    </Switch>}
-  </>
-}
+    </Switch>
+  );
+};
 
 export default App;
