+++ src/routes/confirm-password/index.ts
@@ -1,0 +1,3 @@
+import ConfirmPasswordLayout from './components/ConfirmPasswordLayout';
+
+export default ConfirmPasswordLayout;
